src/requirements.txt
flask==3.0.0
google-cloud-pubsub==2.18.4
gunicorn==21.2.0